Abstract `Btn` component out of `ToggleBtn`
