+++ src/h5web/toolbar/controls/ToggleBtn.tsx
@@ -1,49 +1,13 @@
-import type { AriaAttributes } from 'react';
-import type { IconType } from 'react-icons';
-import styles from '../Toolbar.module.css';
+import Btn, { BtnProps } from './Btn';
 
-interface Props extends AriaAttributes {
-  label: string;
-  icon?: IconType;
-  iconOnly?: boolean;
-  small?: boolean;
-  raised?: boolean;
+interface Props extends Omit<BtnProps, 'onClick'> {
   value: boolean;
   onToggle: () => void;
-  disabled?: boolean;
 }
 
 function ToggleBtn(props: Props) {
-  const {
-    label,
-    icon: Icon,
-    iconOnly,
-    small,
-    raised,
-    value,
-    onToggle,
-    disabled,
-    ...ariaAttrs
-  } = props;
-
-  return (
-    <button
-      className={styles.btn}
-      type="button"
-      onClick={() => onToggle()}
-      disabled={disabled}
-      data-small={small || undefined}
-      data-raised={raised || undefined}
-      aria-label={iconOnly ? label : undefined}
-      aria-pressed={value}
-      {...ariaAttrs}
-    >
-      <span className={styles.btnLike}>
-        {Icon && <Icon className={styles.icon} />}
-        {!iconOnly && <span className={styles.label}>{label}</span>}
-      </span>
-    </button>
-  );
+  const { value, onToggle, ...btnProps } = props;
+  return <Btn {...btnProps} aria-pressed={value} onClick={() => onToggle()} />;
 }
 
 export default ToggleBtn;
